Add dirty mode for sync

+++ carbonate/sync.py
@@ -91,8 +91,9 @@ def heal_metric(source, dest):
             logging.warn("Failed to copy %s! %s" % (dest, e))
 
 
-def run_batch(metrics_to_sync, remote, local_storage, rsync_options):
-    staging_dir = mkdtemp()
+def run_batch(metrics_to_sync, remote, local_storage, rsync_options,
+              remote_ip, dirty):
+    staging_dir = mkdtemp(prefix=remote_ip)
     sync_file = NamedTemporaryFile(delete=False)
 
     metrics_to_heal = []
@@ -123,5 +124,9 @@ def run_batch(metrics_to_sync, remote, local_storage, rsync_options):
     print "    Total time: %ss" % total_time
 
     # Cleanup
-    rmtree(staging_dir)
+    if dirty:
+        print "    dirty mode: left temporary directory %s" % staging_dir
+    else:
+        rmtree(staging_dir)
+
     os.unlink(sync_file.name)
